Encode received text as utf-8 rather than latin1
When utf-8 text is received, the WebsocketTransport was using six.b() to
convert it to bytes. This function assumes the text is latin1. Later the
bytes are decoded back into a string when parse_socketIO_packet_data
calls symmetries.decode_string(). This function assumes the bytes are
utf-8 encoded. Since utf-8 is an almost universal standard it is better to
initially convert using utf-8 instead of latin1.